changed property type on rating of all classes and updated migrations

+++ Server/Migrations/MySQL/20240214112429_InitialCreatePostgreSQL.cs
@@ -40,7 +40,7 @@ namespace LCPCollection.Server.Migrations.MySQL
                         .Annotation("MySql:CharSet", "utf8mb4"),
                     Genre = table.Column<string>(type: "longtext", nullable: true)
                         .Annotation("MySql:CharSet", "utf8mb4"),
-                    Rating = table.Column<double>(type: "double", nullable: true)
+                    Rating = table.Column<float>(type: "float", nullable: true)
                 },
                 constraints: table =>
                 {
@@ -73,7 +73,7 @@ namespace LCPCollection.Server.Migrations.MySQL
                         .Annotation("MySql:CharSet", "utf8mb4"),
                     Genre = table.Column<string>(type: "longtext", nullable: true)
                         .Annotation("MySql:CharSet", "utf8mb4"),
-                    Rating = table.Column<double>(type: "double", nullable: true)
+                    Rating = table.Column<float>(type: "float", nullable: true)
                 },
                 constraints: table =>
                 {
@@ -128,7 +128,7 @@ namespace LCPCollection.Server.Migrations.MySQL
                         .Annotation("MySql:CharSet", "utf8mb4"),
                     Genre = table.Column<string>(type: "longtext", nullable: true)
                         .Annotation("MySql:CharSet", "utf8mb4"),
-                    Rating = table.Column<double>(type: "double", nullable: true)
+                    Rating = table.Column<float>(type: "float", nullable: true)
                 },
                 constraints: table =>
                 {
@@ -161,7 +161,7 @@ namespace LCPCollection.Server.Migrations.MySQL
                         .Annotation("MySql:CharSet", "utf8mb4"),
                     Genre = table.Column<string>(type: "longtext", nullable: true)
                         .Annotation("MySql:CharSet", "utf8mb4"),
-                    Rating = table.Column<double>(type: "double", nullable: true)
+                    Rating = table.Column<float>(type: "float", nullable: true)
                 },
                 constraints: table =>
                 {
@@ -190,7 +190,7 @@ namespace LCPCollection.Server.Migrations.MySQL
                         .Annotation("MySql:CharSet", "utf8mb4"),
                     Platforms = table.Column<string>(type: "longtext", nullable: true)
                         .Annotation("MySql:CharSet", "utf8mb4"),
-                    Rating = table.Column<double>(type: "double", nullable: true),
+                    Rating = table.Column<float>(type: "float", nullable: true),
                     DateCreated = table.Column<DateTime>(type: "datetime(6)", nullable: true),
                     AuthorName = table.Column<string>(type: "longtext", nullable: true)
                         .Annotation("MySql:CharSet", "utf8mb4"),
@@ -228,7 +228,7 @@ namespace LCPCollection.Server.Migrations.MySQL
                         .Annotation("MySql:CharSet", "utf8mb4"),
                     Genre = table.Column<string>(type: "longtext", nullable: true)
                         .Annotation("MySql:CharSet", "utf8mb4"),
-                    Rating = table.Column<double>(type: "double", nullable: true)
+                    Rating = table.Column<float>(type: "float", nullable: true)
                 },
                 constraints: table =>
                 {
@@ -256,7 +256,7 @@ namespace LCPCollection.Server.Migrations.MySQL
                         .Annotation("MySql:CharSet", "utf8mb4"),
                     BrowsersName = table.Column<string>(type: "longtext", nullable: true)
                         .Annotation("MySql:CharSet", "utf8mb4"),
-                    Rating = table.Column<double>(type: "double", nullable: true),
+                    Rating = table.Column<float>(type: "float", nullable: true),
                     DateCreated = table.Column<DateTime>(type: "datetime(6)", nullable: true),
                     AuthorName = table.Column<string>(type: "longtext", nullable: true)
                         .Annotation("MySql:CharSet", "utf8mb4"),
@@ -272,7 +272,7 @@ namespace LCPCollection.Server.Migrations.MySQL
             migrationBuilder.InsertData(
                 table: "Animes",
                 columns: new[] { "Id", "Companies", "CoverUrl", "Description", "Genre", "ImageUrl", "Platforms", "Publishers", "Rating", "ReleaseDate", "Title", "TrailerUrl" },
-                values: new object[] { 1, "Toei Animation", "covers/db.jpg", "Dragon Ball", "Animation", "images/db.jpg", "TV", "Toei Animation", 9.0, new DateTime(1986, 2, 26, 0, 0, 0, 0, DateTimeKind.Utc), "Dragon Ball", "https://www.youtube.com/watch?v=gqIEgmqljM8" });
+                values: new object[] { 1, "Toei Animation", "covers/db.jpg", "Dragon Ball", "Animation", "images/db.jpg", "TV", "Toei Animation", 9f, new DateTime(1986, 2, 26, 0, 0, 0, 0, DateTimeKind.Utc), "Dragon Ball", "https://www.youtube.com/watch?v=gqIEgmqljM8" });
         }
 
         /// <inheritdoc />
