+++ Shared/Classes/Softwares.cs
@@ -30,7 +30,7 @@ public class Softwares {
     [DataType("text")]
     public string? Platforms { get; set; }
 
-    public double? Rating { get; set; }
+    public float? Rating { get; set; }
 
     [DataType("datetime")]
     public DateTime? DateCreated { get; set; }
